fix: sjekk om instans finnes før vi prøver å slette

+++ pkg/compute/compute.go
@@ -53,9 +53,17 @@ func (c Client) Delete(ctx context.Context, email string, log logger.Logger) boo
 		return true
 	}
 
-	if err := c.deleteComputeInstanceFromGCP(ctx, instance.Name); err != nil {
-		log.WithError(err).Error("failed deleting compute instance from GCP")
-		return false
+	exists, err := c.computeInstanceExistsInGCP(instance.Name)
+	if err != nil {
+		log.WithError(err).Error("failed checking if compute instance exists in GCP")
+		return true
+	}
+
+	if exists {
+		if err := c.deleteComputeInstanceFromGCP(ctx, instance.Name); err != nil {
+			log.WithError(err).Error("failed deleting compute instance from GCP")
+			return true
+		}
 	}
 
 	if err = c.repo.ComputeInstanceDelete(ctx, email); err != nil {
